Added users and user authentication

+++ client/src/components/ComplaintDetails.js
@@ -1,3 +1,4 @@
+import { useAuthContext } from '../hooks/useAuthContext'
 import { useComplaintsContext } from '../hooks/useComplaintsContext'
 
 // date fns
@@ -5,10 +6,16 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const ComplaintDetails = ({ complaint }) => {
   const { dispatch } = useComplaintsContext()
-
+  const { user } = useAuthContext()
   const handleClick = async () => {
+    if(!user){
+      return 
+    }
     const response = await fetch('/api/complaints/' + complaint._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers:{
+        'Authorization':`Bearer ${user.token}`
+      }
     })
     const json = await response.json()
